+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 6.19.1
+
+### Core
+
+- adding _evt to arc rotate's pointer input - by [RaananW](https://github.com/RaananW) ([#14241](https://github.com/BabylonJS/Babylon.js/pull/14241))
+- SolidParticleSystem: Add uvKind option to the digest method - by [Popov72](https://github.com/Popov72) ([#14237](https://github.com/BabylonJS/Babylon.js/pull/14237))
+
+### Node Editor
+
+- fix alignment issue in NME - [_Bug Fix_] by [carolhmj](https://github.com/carolhmj) ([#14240](https://github.com/BabylonJS/Babylon.js/pull/14240))
+
 ## 6.19.0
 
 ### Core